Move codegen_stmt_debuginfo to debuginfo.rs
